Fix: 카테고리 설명 모달창 수정
