+++ src/pages/Collect/Challenge/PublicChallenge/PublicChallenge.jsx
@@ -1,20 +1,18 @@
 import React, { useState } from "react";
 import arrowRight from "../../../../assets/Navigation/arrowRight.svg";
-import arrowUp from "../../../../assets/Navigation/arrowUp.svg";
-import arrowDown from "../../../../assets/Navigation/arrowDown.svg";
 import styles from "./PublicChallenge.module.css";
 import chatPlusBtn from "../../../../assets/Btn/chatPlusBtn.svg";
 import { Link, useNavigate } from "react-router-dom";
 import JoinChallenge from "./JoinChallenge";
 import ChallengeCard from "../../../../components/ChallengeCard/ChallengeCard";
 import SearchBar from "../../../../components/SearchBar/SearchBar";
+import Dropdown from "../../../../components/Dropdown/Dropdown";
 
 const PublicChallenge = ({ allData }) => {
   const navigate = useNavigate();
   const noJoinChallenge = allData.filter((data) => data.isJoined == false);
   const joinChallenge = allData.filter((data) => data.isJoined == true);
   const [sortChallenge, setSortChallenge] = useState(noJoinChallenge);
-  const [isDropDown, setIsDropDown] = useState(false);
   const sortType = [
     "인기순",
     "최신순",
@@ -22,33 +20,6 @@ const PublicChallenge = ({ allData }) => {
     "코인많은순",
     "코인적은순",
   ];
-  const [sortName, setSortName] = useState(sortType[0]);
-
-  const handleSortChallenge = (name) => {
-    setSortName(name);
-    const compare = (a, b) => {
-      let date1, date2;
-      switch (name) {
-        case "인기순":
-          return b.people - a.people;
-        case "최신순":
-          date1 = new Date(a.startDate);
-          date2 = new Date(b.startDate);
-          return date1 > date2 ? 1 : -1;
-        case "종료임박순":
-          date1 = new Date(a.endDate);
-          date2 = new Date(b.endDate);
-          return date1 > date2 ? 1 : -1;
-        case "코인많은순":
-          return b.coin - a.coin;
-        case "코인적은순":
-          return a.coin - b.coin;
-        default:
-          break;
-      }
-    };
-    setSortChallenge([...noJoinChallenge].sort(compare));
-  };
 
   return (
     <>
@@ -102,26 +73,11 @@ const PublicChallenge = ({ allData }) => {
         )}
 
         <h3>이런 챌린지는 어떠세요?</h3>
-        <div className={styles.dropDown}>
-          <button type="button" onClick={() => setIsDropDown(!isDropDown)}>
-            <p>{sortName}</p>
-            <img
-              src={isDropDown ? arrowUp : arrowDown}
-              alt="정렬 버튼 아이콘"
-            />
-            {isDropDown && (
-              <ul>
-                {sortType
-                  .filter((name) => name !== sortName)
-                  .map((name) => (
-                    <li key={name} onClick={() => handleSortChallenge(name)}>
-                      {name}
-                    </li>
-                  ))}
-              </ul>
-            )}
-          </button>
-        </div>
+        <Dropdown
+          sortArr={sortChallenge}
+          setSortArr={setSortChallenge}
+          sortType={sortType}
+        />
         <div className={styles.publicChallengeWrapper}>
           {sortChallenge.map((item) => (
             <ChallengeCard
