+++ src/pages/MyRecord/ShowMyChallenge.jsx
@@ -6,35 +6,12 @@ import question from "../../assets/Content/question.svg";
 import arrowUp from "../../assets/Navigation/arrowUp.svg";
 import arrowDown from "../../assets/Navigation/arrowDown.svg";
 import Header from "../../components/Header/Header";
+import Dropdown from "../../components/Dropdown/Dropdown";
 
 const ShowMyChallenge = () => {
   const [sortChallenge, setSortChallenge] = useState(dummyData);
-  const [isDropDown, setIsDropDown] = useState(false);
   const sortType = ["최신순", "오래된 순", "코인순"];
-  const [sortName, setSortName] = useState(sortType[0]);
   const [modalOpen, setModalOpen] = useState(false);
-
-  const handleSortChallenge = (name) => {
-    setSortName(name);
-    const compare = (a, b) => {
-      let date1, date2;
-      switch (name) {
-        case "최신순":
-          date1 = new Date(a.startDate);
-          date2 = new Date(b.startDate);
-          return date1 > date2 ? 1 : -1;
-        case "오래된 순":
-          date1 = new Date(a.startDate);
-          date2 = new Date(b.startDate);
-          return date1 < date2 ? 1 : -1;
-        case "코인순":
-          return b.coin - a.coin;
-        default:
-          break;
-      }
-    };
-    setSortChallenge([...dummyData].sort(compare));
-  };
 
   return (
     <>
@@ -49,35 +26,20 @@ const ShowMyChallenge = () => {
               onClick={() => setModalOpen(true)}
             />
           </div>
-          {modalOpen && (
-            <DescModal modalOpen={modalOpen} setModalOpen={setModalOpen} />
-          )}
         </div>
-        <div className={styles.dropDown}>
-          <button type="button" onClick={() => setIsDropDown(!isDropDown)}>
-            <p>{sortName}</p>
-            <img
-              src={isDropDown ? arrowUp : arrowDown}
-              alt="정렬 버튼 아이콘"
-            />
-            {isDropDown && (
-              <ul>
-                {sortType
-                  .filter((name) => name !== sortName)
-                  .map((name) => (
-                    <li key={name} onClick={() => handleSortChallenge(name)}>
-                      {name}
-                    </li>
-                  ))}
-              </ul>
-            )}
-          </button>
-        </div>
+        <Dropdown
+          sortArr={sortChallenge}
+          setSortArr={setSortChallenge}
+          sortType={sortType}
+        />
         <div className={styles.challengeRecordWrapper}>
           {sortChallenge.map((item) => (
             <MyChallengeRecord id={item.id} item={item} isConsData={true} />
           ))}
         </div>
+        {modalOpen && (
+          <DescModal modalOpen={modalOpen} setModalOpen={setModalOpen} />
+        )}
       </div>
     </>
   );
