+++ src/pages/MyRecord/ShowMyChallenge.module.css
@@ -29,63 +29,6 @@
   line-height: normal;
 }
 
-.dropDown {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  align-items: flex-end;
-  margin-bottom: 15px;
-  position: relative;
-}
-ul {
-  display: flex;
-  box-sizing: border-box;
-  flex-direction: column;
-  position: absolute;
-  top: 26px;
-  right: 0;
-  width: 90px;
-  align-items: center;
-  padding: 0px 8px;
-  gap: 4px;
-  border-radius: 6px;
-  background: var(--Grey-70, #f1f1f1);
-  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.25);
-}
-li {
-  color: var(--Grey-800, #2b2b2b);
-  text-align: center;
-  padding: 2px 0;
-  /* md/12px */
-  font-family: Pretendard;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-.dropDown button {
-  display: flex;
-  box-sizing: content-box;
-  width: 90px;
-  height: 24px;
-  justify-content: flex-end;
-  align-items: center;
-  border-radius: 8px;
-  background: var(--Grey-70, #f1f1f1);
-}
-.dropDown button p {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  padding-left: 5px;
-  color: var(--Grey-800, #2b2b2b);
-  text-align: center;
-  font-family: Pretendard;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
 .challengeRecordWrapper {
   display: flex;
   width: 100%;
